requirements.txt
flask
gunicorn
python-docx
docxtpl
pypandoc
lxml
flask-cors


